# storage/api-docs/api-docs.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "PBL SIPASTI API",
        "description": "PBL SIPASTI API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/kelola-role": {
            "post": {
                "tags": [
                    "Kelola Role"
                ],
                "summary": "Get list of roles",
                "description": "Get list of roles with pagination and search",
                "operationId": "4d125a8f5bb53bdc639274ed1d3ef692",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/KelolaRoleRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Data berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "message": "Data berhasil diambil",
                                    "data": "data"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": false,
                                    "message": "Unauthorized"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": false,
                                    "message": "Internal Server Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "cookieAuth": [
                            "jwtToken"
                        ]
                    }
                ]
            }
        },
        "/api/kelola-role/{id}": {
            "get": {
                "tags": [
                    "Kelola Role"
                ],
                "summary": "Get role detail",
                "description": "Get role detail",
                "operationId": "d21e0414790de30782f9f4e892525f16",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Role ID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "message": "Data berhasil diambil",
                                    "data": "data"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": false,
                                    "message": "Unauthorized"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": false,
                                    "message": "Internal Server Error"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": false,
                                    "message": "User not found"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "cookieAuth": [
                            "jwtToken"
                        ]
                    }
                ]
            },
            "put": {
                "tags": [
                    "Kelola Role"
                ],
                "summary": "Update role",
                "description": "Update role",
                "operationId": "ae9d8c8064ee401c4fa773d19a8e3b7e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Role ID",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/KelolaRoleUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Data berhasil diubah"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    },
                    "404": {
                        "description": "User not found"
                    }
                },
                "security": [
                    {
                        "cookieAuth": [
                            "jwtToken"
                        ]
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Kelola Role"
                ],
                "summary": "Delete role",
                "description": "Delete role",
                "operationId": "09a95fad9f6a8a76a102f8bd42c83c42",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User ID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data berhasil dihapus"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    },
                    "404": {
                        "description": "User not found"
                    }
                },
                "security": [
                    {
                        "cookieAuth": [
                            "jwtToken"
                        ]
                    }
                ]
            }
        },
        "/api/kelola-role/create": {
            "post": {
                "tags": [
                    "Kelola Role"
                ],
                "summary": "Create role",
                "description": "Create role",
                "operationId": "22a7f8d92a3efcf10448debeb7e6cca3",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/KelolaRoleCreateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Data berhasil ditambahkan",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "message": "Data berhasil ditambahkan",
                                    "data": "data"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": false,
                                    "message": "Unauthorized"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": false,
                                    "message": "Internal Server Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "cookieAuth": [
                            "jwtToken"
                        ]
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "KelolaRoleRequest": {
                "properties": {
                    "search": {
                        "type": "string",
                        "example": ""
                    }
                },
                "type": "object"
            },
            "KelolaRoleCreateRequest": {
                "properties": {
                    "role_nama": {
                        "type": "string",
                        "example": ""
                    }
                },
                "type": "object"
            },
            "KelolaRoleUpdateRequest": {
                "properties": {
                    "role_nama": {
                        "type": "string",
                        "example": ""
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}